cv_backend/requirements.txt
asgiref==3.11.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
pillow==12.1.1
psycopg2-binary
sqlparse==0.5.5
gunicorn==21.2.0
djangorestframework-simplejwt==5.5.1
whitenoise==6.11.0
django-ratelimit==4.1.0
django-redis==6.0.0
redis
django-redis